功能提交

+++ src/main/java/com/cehome/apimanager/service/impl/AmProjectServiceImpl.java
@@ -1,19 +1,22 @@
 package com.cehome.apimanager.service.impl;
 
-import java.util.Date;
-import java.util.List;
-
+import com.cehome.apimanager.common.Page;
+import com.cehome.apimanager.dao.AmProjectDao;
+import com.cehome.apimanager.exception.BizValidationException;
+import com.cehome.apimanager.model.dto.AmModuleQueryReqDto;
+import com.cehome.apimanager.model.dto.AmProjectQueryReqDto;
+import com.cehome.apimanager.model.dto.AmProjectReqDto;
+import com.cehome.apimanager.model.dto.AmProjectResDto;
+import com.cehome.apimanager.model.po.AmModule;
+import com.cehome.apimanager.model.po.AmProject;
+import com.cehome.apimanager.service.IAmModuleService;
+import com.cehome.apimanager.service.IAmProjectService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.cehome.apimanager.common.Page;
-import com.cehome.apimanager.dao.AmProjectDao;
-import com.cehome.apimanager.model.dto.AmProjectQueryReqDto;
-import com.cehome.apimanager.model.dto.AmProjectReqDto;
-import com.cehome.apimanager.model.dto.AmProjectResDto;
-import com.cehome.apimanager.model.po.AmProject;
-import com.cehome.apimanager.service.IAmProjectService;
+import java.util.Date;
+import java.util.List;
 
 /**
  * 项目业务接口实现
@@ -26,6 +29,9 @@ public class AmProjectServiceImpl implements IAmProjectService {
 
 	@Autowired
 	private AmProjectDao projectDao;
+
+	@Autowired
+	private IAmModuleService moduleService;
 
 	@Override
 	public void add(AmProjectReqDto dto) {
@@ -53,6 +59,12 @@ public class AmProjectServiceImpl implements IAmProjectService {
 
 	@Override
 	public void delete(AmProjectReqDto dto) {
+		AmModuleQueryReqDto moduleQueryReqDto = new AmModuleQueryReqDto();
+		moduleQueryReqDto.setProjectId(dto.getId());
+		List<AmModule> moduleList = moduleService.list(moduleQueryReqDto);
+		if(moduleList != null && !moduleList.isEmpty()){
+			throw new BizValidationException("项目下存在其他模块，不能删除！");
+		}
 		projectDao.delete(dto.getId());
 	}
